+++ src/Commands/InstallCommand.php
@@ -2,7 +2,8 @@
 
 namespace Nuntius\Commands;
 
-use MongoDB\Client;
+use Nuntius\Db\MongoDB\MongoDBTraitInstallation;
+use Nuntius\Db\RethinkDB\RethinkDbTraitInstallation;
 use Nuntius\Nuntius;
 use Symfony\Component\Console\Command\Command;
 use Symfony\Component\Console\Input\InputInterface;
@@ -15,7 +16,9 @@ use Symfony\Component\Yaml\Yaml;
 /**
  * CLI command to install the bot.
  */
-class InstallCommand extends Command  {
+class InstallCommand extends Command {
+
+  use RethinkDbTraitInstallation, MongoDbTraitInstallation;
 
   /**
    * {@inheritdoc}
@@ -117,86 +120,6 @@ class InstallCommand extends Command  {
     $fs->dumpFile(__DIR__ . '/../../settings/credentials.local.yml', $yml_content);
 
     return $settings;
-  }
-
-  /**
-   * Set the settings for RethinkDB.
-   *
-   * @param $settings
-   *  The settings variable.
-   */
-  protected function setDbRethinkdbSettings(&$settings, SymfonyStyle $io) {
-    $settings['rethinkdb']['host'] = $io->ask('Enter the address of the DB', 'localhost');
-    $settings['rethinkdb']['port'] = $io->ask('Enter the port address of the DB', 28015);
-    $settings['rethinkdb']['db'] = $io->ask('Enter the DB name', 'nuntius');
-    $settings['rethinkdb']['api_key'] = $io->ask('Enter the API key', 'none');
-    $settings['rethinkdb']['timeout'] = $io->ask('Enter the timeout for the DB connection', 30);
-
-    if ($settings['rethinkdb']['api_key'] == 'none') {
-      $settings['rethinkdb']['api_key'] = '';
-    }
-  }
-
-  /**
-   * Check the connection.
-   *
-   * @param $settings
-   * @param SymfonyStyle $io
-   */
-  protected function checkDbRethinkdbSettings(&$settings, SymfonyStyle $io) {
-    try {
-      @\r\connect($settings['rethinkdb']['host'], $settings['rethinkdb']['port'], $settings['rethinkdb']['db'], $settings['rethinkdb']['api_key'], $settings['rethinkdb']['timeout']);
-    } catch (\Exception $e) {
-      $io->error("Hmm.. It seems there is an error: " . $e->getMessage() . ". Let's start again.");
-      return FALSE;
-    }
-
-    return TRUE;
-  }
-
-  /**
-   * Set the settings for MongoDB.
-   *
-   * @param $settings
-   * @param SymfonyStyle $io
-   */
-  protected function setDbMongodbSettings(&$settings, SymfonyStyle $io) {
-    $settings['mongodb']['uri'] = $io->ask('Enter the URI of the server', 'mongodb://127.0.0.1/');
-    $settings['mongodb']['db'] = $io->ask('Enter the DB name', 'nuntius');
-
-    if ($username = $io->ask('Enter the username', FALSE)) {
-      $settings['mongodb']['username'] = $username;
-    }
-
-    if ($password = $io->ask('Enter the password', FALSE)) {
-      $settings['mongodb']['password'] = $password;
-    }
-  }
-
-  /**
-   * Check the connection for MongoDB.
-   *
-   * @param $settings
-   * @param SymfonyStyle $io
-   */
-  protected function checkDbMongodbSettings(&$settings, SymfonyStyle $io) {
-    $options = [];
-
-    if (!empty($settings['mongodb']['username']) && !empty($settings['mongodb']['password'])) {
-      // Setting up username and password.
-      $options['username'] = $settings['mongodb']['username'];
-      $options['password'] = $settings['mongodb']['password'];
-    }
-
-    try {
-      $collection = new Client($settings['mongodb']['uri'], $options);
-      $collection->listDatabases();
-    } catch (\Exception $e) {
-      $io->error("Hmm.. It seems there is an error: " . $e->getMessage() . ". Let's start again.");
-      return FALSE;
-    }
-
-    return TRUE;
   }
 
 }
